Use branch commit ref in recipe checksum.
If the tag specified is actually a branch, resolve the commit and use it to
calculate the recipe hash. We need to do this only for branches, since tags we
consider immutable in any case. Notice that if we decide to do it also for tags
adding --tags is not enough, since ls-remote would return the hash of the tag
object itself, not the commit.  You can get the actual hash for the commit by
using tag^{}.