Improve logging to associate it with a given build.
When using debug output, we now print out the "main package" and the hash used
for it. This will allow us to aggregate different builds of the same main
package, e.g. when building for different architectures or with different
version of some external (e.g., geant 4). This should simplify comparisons
and parsing with logstash.