Use tarball symlinks to determine revision.
This uses the tarball symlinks, rather than the actual installation directory,
to derive the installation revision. This is done so that we can actually sync
the symlinks from server and ensure we have a consistent setup, not only for
the hash based storage, but also for the tarball friendly names.