Handle corner case with overlapping versions. Fix alisw/alidist#72.
There is a corner case when two packages have an overlapping version
which includes a -, say alice3 and alice3-xrd. If the latter was already
uploaded, the former will still think its revision is a possible
candidate, but will then fail to match later on, which is correct.
However if there is no other tarball available, we will end up crashing,
rather than simply assigning "1" as a revision.

This handles the problem by simply detecting the fact there is no
revision candidates and therefore picking 1.