Use develPrefix as version for devel packages
This fixes #212: in case two different `develPrefix`es are used for two
builds, build directories are different (as the hashes will differ) and
installation directories will be different too (as the version is used in the
directory name).

To prevent people currently using `ali-master` as `develPrefix` to have
everything rebuilt because of the hash change, this fix does not apply in that
case.