Avoid spurios messages in aliDoctor
This should obviate the issue with prefer_system_check using brew (on
Mac) to looks for system packages. On linux this fails because brew is
not there creating confusion, while being harmless.

This fixes the problem by replacing "brew" with "true" in case brew is
not found in path.

Fixes alisw/alidist#296.