Improve and refactor aliDoctor
The package list creation logic has been abstracted out of aliBuild and
now powers aliDoctor. This allows us to limit the check to the actual
packages being built and not all of them. Moreover it also allows for
checking the system requirements.

Output is also prettified and human readable.

Fixes #267.