Add Python3 support
This should make aliBuild work when using python3, at the cost of
dropping support for python <2.5.

In particular this does the following:

- Use "as" for Exceptions catching
- Use of print_function
- Handle different imports in python2 and python3
- Use str in place of basestring
- Use items() rather than iteritems()
- Use absolute imports
- Align raw_input usage

Moreover this drops the internal pyyaml, the rationale being that now
that we have proper pip setup, we can simply tell people to install it
using it. If one wants to develop, he / she can probably already install
it by hand.