Protect again aliBuild in alienv cornercase
aliBuild was not protecting against previously setup environment under
some circunstances. This should make sure the build environment will
always override the one previously set.

Notice there might still be leaks of the user environment into the build
one, but to have a definitive solution for this we would need to move to
rely on modulefiles only.