Fix issue with wrong order of environment variables
Reading the content of the header section of any recipe results in an arbitrary order of the variables
defined in the env section. Often this order is importatnt and should be the same as defined in the
env section of the header.
Add functionality which loads the content of the header section into a ordered dictionary such that
the order of the variables is preserved.