Introduce the concept of build family
This introduces the concept of build family. Basically until now
only the development packages were getting a latest-<devel-prefix>
link, while any of their descendents did not. This leads to issue
when cleaning up as described in:

    https://github.com/alisw/alibuild/issues/325

With this change we introduce a new concept, the build family, which is
common to all packages which are involved in the build of a development
package or which use a particular default. It allows us to create a
latest-<build-family> link for each package involved in the build.
<build-family> will be either:

- <devel-prefix>-<defaults>
- <defaults>

depending on whether the package is a devel package or not.