Introduce valid_defaults metadata
Some packages can only be compiled using a given default. Most notably, the O2
package can only be built with `--defaults o2` or `--defaults o2-daq`. It's now
possible to express that by adding a valid_defaults entry in the YAML preable
which contains the list of valid defaults. Notice this will be checked for all
the packages to be built.