Fix relocation for CMake files
Technically, this fix does not just change the relocation for the case
of CMake files. But, excluding the semi-colon from the set of characters
expected to be used in directory names seems fairly generic though. We
already exclude spaces, which are far more likely to be used in
directory names.

Fixes #508