Make sure new modules works in a mode we support
This fixes a critical issue with the return codes, among other things. Old
`modulecmd` returned 0 when some dependencies were not found, and this allowed
us to perform tricks such as loading the correct GCC-Toolchain on user laptops,
or being tolerant with malformed Modulefiles (allowing people to keep working
in spite of the various warnings saying that certain packages could not be
found).

The new one aborts execution in case a dependency is not found, breaking our
user space without notice. We therefore force `modulecmd` to run in
compatibility mode to know what we are dealing with.

Note that this variable also fixes the `MODULEPATH` problem with empty path
strings, and macOS warnings for `DYLD_*_modshare` variables, but it's better to
keep those fixes for the future, in case we will be forced to lift the
compatibility mode.