Separate local-only builds from remote revisions
This prevents interference between the remote store and local builds in
the case where another machine builds a package and uploads it to the
remote store with a revision number that conflicts with a local package,
in case that package wasn't uploaded to the remote store. In that case,
local revision numbers become a mess, and packages could start depending
on the wrong revision numbers of dependencies.

It also prevents builds intended only for local use from being uploaded
to the remote. This is important as local-only builds might otherwise be
uploaded to the remote much later, when some of their dependencies might
conflict with new ones on the remote, and would be renamed. In that
case, we currently don't update the main package, so it would depend on
(and e.g. try to source scripts from) an incorrect revision number of
its dependency.

With these changes, if running aliBuild with a read-write remote when
some local revisions are left over in the local store from a previous
run, then those revisions are not uploaded, and may be deleted from the
local store. Running aliBuild repeatedly with a read-only remote
preserves local revisions.