Don't depend on local revisions when uploading
When packages can be uploaded, they should not be built to depend on
local-only revisions of dependencies. This isn't a problem currently as
the remote store handlers delete local-only revisions on sync, but might
be in the future.

Also allow the use of remote revisions when building locally, to reduce
needless rebuilds.