Fetch the tag from existing git repos if we can't checkout
This fetches the tag we want to checkout for an existing git repo if it can't be
checked out, then retries checking out the desired tag. This only results in an
extra network request per repo (which might mean an extra password prompt) if
the tag doesn't already exist, which happens if the tag was made after the repo
in SOURCES was first cloned.

This doesn't catch cases where a tag moves (i.e. changes from one commit to
another), or where $GIT_TAG is a branch, not a tag.