Calculate separate hashes for local and remote revisions
Local revisions get a different hash (with the same information as remote
revisions, but an extra "local" string in the input to the hash). This is to
avoid conflicts where a remote might gain a package with the same hash as a
locally-built package; in that case, the two would conflict. With this patch,
the two revisions of the same package should be able to coexist.

Syncing from and to remote stores can always use the remote-specific hashes.
Once we have assigned a revision to the package (or reused an existing one), we
know which hash (remote or local) to assign and use for the build.

This commit also extracts the hash calculation into a separate function for
cleanliness.

The same hashes are still generated for the same specs, so existing tarballs on
the remote store are used.

A bit of reorganisation was necessary to move the hash calculation into the main
build loop, as we need to know our dependencies' hashes at the time we calculate
our own hash, but to know their hashes, they must have first been assigned a
revision number (so that we know whether to use the local or remote hash).