Don't redownload existing symlinks
We shouldn't be producing conflicting symlinks anyway, and if we do, it's
probably better if we just keep our local one.

This should reduce the remote sync time significantly on subsequent runs of
aliBuild (though we'll still have to download everything on the first run).

On slc7, when we don't have to rebuild anything, this change takes the runtime
of `aliBuild build O2 --defaults o2` from 5 minutes to 16 seconds.