Fix aliBuild not detecting that a dev package needs rebuilding (#671)
For some reason, this was broken recently, such that after changing a
development package's sources, it would need two aliBuild invocations to
actually be rebuilt. This patch should fix this (and testing with AliPhysics, it
does).

I'm slightly worried this might introduce unnecessary rebuilds, but in my
testing I haven't come across any (yet?).

I'm still not sure why this bug has suddenly appeared as this particular code
hasn't been touched in a while. Perhaps it has to do with the recent reshuffling
of the hashing logic.