Improve ability to filter requires
This extends the ability to filter requires, currently possible
only on a per architecture basis, to defaults and a newly introduced
"target".
The old syntax:

  requires:
    zlib: osx_x86-64   # Only on macOS

is still valid and equivalent to:

  requires:
    zlib: arch=osx_x86-64   # Only on macOS

One can now also do:

  requires:
    zlib: defaults=o2       # Only when compiling O2

or:

  requires:
    zlib: target=analysis        # Only when invoking `alibuild build O2:analysis`

Multiple filters can now be specified via comma separation,
and will be logically ANDed together.