Use a manifest file to speed up symlink downloads
With this commit, aliBuild looks for a <package>.manifest file listing symlink
names and targets, in order to avoid having to download many small files.

The manifest file is a simple text file, so symlink names and targets must not
contain newlines, and symlink names must not contain tab characters.

This change is backwards-compatible: the many individual pseudo-symlinks can
stay; they will be simply ignored (and not downloaded) if they appear in the
manifest. Old aliBuild versions will ignore the manifest and can continue to use
the individual pseudo-symlink files.