Consider all possible hashes when building tag
Store alternative hashes for a spec, for each other git reference we could've
used to point to the commit we're actually building. This means that we can
e.g. reuse tarballs built from commit abcde when building tag foo, if foo tags
the commit abcde.

This currently works both ways -- if we're building a commit, look for tags
pointing to it, and if we're building a tag, look for the commit hash and other
tags pointing to the same commit. This makes the assumption that tags don't
move around!

Tested using python3 on slc7 and slc8 for each remote store (rsync, S3, boto3,
HTTP).