Prefer remote revisions to existing local revisions (#718)
If we have both a local and a remote tarball that both match the hashes we want,
reuse the remote tarball instead of the first one we find. This allows us to
potentially reuse remote (or local) tarballs for later packages. If we find
multiple reusable tarballs, prefer the one with the "primary" hash.