Use host network for Docker builds, and no network for system checks
System checks shouldn't need to use the network, so disable it in Docker builds
for a speed-up.

For regular builds, we can just use the host's network stack as we don't need
any particular network features from Docker, and this way it's faster to set up
and tear down containers.