Use a connection pool for HTTP remote stores (#741)
* Use a connection pool for HTTP remote stores

This way, we only connect to a HTTP remote once, and then keep reusing this
connection for each request. This should give a speed-up, especially for the
many small requests when listing directories.

* Fix python2 mocking of alibuild_helpers.git

* Create a separate Session per package

This stops each connection from being kept open for very long; instead, each
connection is only used for a few seconds (but potentially many requests) and
then closed.