Rebuild later packages if dev package changes (#739)
If a development package has committed or uncommitted changes, any packages
depending on it should be rebuilt with the new code. This patch implements this.

This is done by incorporating the `devel_hash` of every dependency (if present)
in a package's overall hash, so that if a dependency changes, we rebuild
packages that depend on it. A dev package's hash itself doesn't change, however,
so we keep rebuilding it in the same build area.