Add ability to disable / enable packages by architecture
This allows us to define that a given package can only be
built on certain architectures. This is done via the architecture
tag, e.g.:

architecture: slc7 # build on slc7 only
architecture: (?!osx) # Do not build on macOS
architecture: (slc7|slc6) # Only slc6 and slc7

This avoids having to specify in every dependent of the package
the conditional requirement string. It also has the advantage
that when a package is disabled, but still present on the
system because