Document ways to cache git credentials
While aliBuild doesn't fail as badly when presenting git credential prompts any
more, it can still be annoying to be prompted often for the same password.

To alleviate this, document ways to cache or eliminate passwords that must be
typed by the user. aliBuild should not do this automatically as it interferes
with the user's configuration.