Allow users to input git credentials without prompts overwriting each other (#775)
* Prevent git password prompts from interfering with each other

If a git repository cannot be accessed in parallel, try it sequentially while
allowing the user to input their credentials. This fixes the issue where
multiple git processes running in parallel would all present a username and
password prompt, overwriting each other.

Avoiding password prompts entirely should be out of scope for aliBuild, as that
requires changing the user's git configuration. Instead, the documentation
should specify how to cache passwords or use SSH auth instead.

* Document ways to cache git credentials

While aliBuild doesn't fail as badly when presenting git credential prompts any
more, it can still be annoying to be prompted often for the same password.

To alleviate this, document ways to cache or eliminate passwords that must be
typed by the user. aliBuild should not do this automatically as it interferes
with the user's configuration.

* Link to troubleshooting docs in aliBuild message

This should make ways to fix the issue of users being prompted for their
password too often easier to find.

* Add unittest covering git wrapper function

* Fix git unittest

In GitHub CI, we have access to the alisw org, so no password prompt is shown.

Instead, just check for the error type, and use another private repo to check
for access restrictions.