Fix tarball unpacking (#778)
Only create a new tarball if we actually need it, and copy in parallel with
creating the tarball as for the latter, CPU usage is the bottleneck without
pigz.

Use pigz only for compression. For decompression, pigz provides very little
speedup, according to its documentation.

* Preserve hardlinks when installing packages

Previously, we round-tripped files through tar to copy them from INSTALLROOT to
their final place. This preserved hardlinks, as that's tar's default behaviour.
Make sure rsync does the same, so we don't break things.