Show helpful error message in case of name clashes
If the user has a local directory named the same as an alidist package, that is
interpreted as a development package (i.e. source code checkout) by aliBuild.
However, if the package has no `source:` key in alidist, then this would not
work. Instead of failing with a Python exception in that case, show a more
helpful error message to the user.