[O2-3609] Limit GitHub workflow permissions (#796)
* [O2-3609] Limit GitHub workflow permissions

* Remove unsupported Python 2 test

The setup-python workflow doesn't support Python 2.7 any more, and getting it
to work any other way is very fiddle. We don't use Python 2 any more anyway --
the only use case would be to reproduce ancient builds on slc6, in which case
we can use alibuild<=1.14.4.

* Reduce workflow permissions further

We don't need `contents: read` as the repo contents are public.