Replace `which` with POSIX-compatible alternative
Due to reported problems using `alienv` in an Ubuntu container with some
contamination from the CentOS 7 host environment, use `command -v` to find
executable locations instead.

`command -v` is defined by POSIX and is a shell builtin in bash, so there
should be no danger of environment contamination.

An alternative would be `type`, but its output is not in the correct format,
and its exit code is not standardised by POSIX, so it may be unreliable as an
existence check for a given executable.