Generate init.sh in one place
This reduces the duplication in the init.sh generation code significantly.

Instead of generating shell commands that generate init.sh, scattered all over
the place, generate the contents of init.sh in one place and just write them
into the file when applicable.

Once created, load the generated file instead of generating the same shell
code twice at different escaping levels.

All of this reduces the potential for errors a lot, since different places in
the code don't need to be kept in sync when generating and re-generating the
file.

Add tests that make sure the generated file contents are vaguely sensible, and
that clobbering the file during the build (as the AliEn-Runtime package does)
does not mean it is mangled during subsequent builds.