Use dieOnError in doBuild
It's a bit simpler to set the exit code this way. With this change, unit tests
also don't need to check the exit code anymore.

This will also make it easier in future to separate out parts of doBuild into
other functions, since we wouldn't have to pass exit codes around.