Save more data in `sw/MIRROR` using `--filter=tree:0`
If git supports it, prefer the `--filter=tree:0` option to
`--filter=blob:none`, but fall back to the latter if the former is not
supported.

If using a Docker container, use neither (as before), since we don't know
whether the git in the container can read the repo we produce outside the
container.