Split fetching from remote into tarball and symlink fetching
This allows us to delay fetching tarballs, and avoid it completely if we
already have the package installed.

Also get rid of a pointless parameter (the package name is already set in the
spec).

To avoid re-fetching symlinks on every run, abort if we already have a
matching one for any of our package's hashes.

Previously, symlinks would not be re-fetched when a tarball was available, but
now we might not have a tarball, even if the package in question is already
installed.