Clone SOURCES repos outside container
This lets us always use speed-up options like --filter, even when we're
building inside a container.

We can also easily start using sapling for this in future by implementing the
new SCM methods.

This also helps with some annoying authentication-related issues, if running
inside --docker, such as in CI.

Fixes <https://its.cern.ch/jira/browse/O2-2439>.

When this is merged, we can also get rid of the `GIT_*` environment variables
used in CI, and stop installing an especially-new Git version in our
containers (since that was needed to understand the `GIT_*` variables).