Standardize on pyproject.toml
This way all installations are PEP517 compatible.

The setuptools scripts is deprecated, but for now it's the best way of
migrating forward.

Taking the chance to do this to fix this issue:

```
ERROR: setuptools==53.0.0 is used in combination with setuptools-scm>=8.x

Your build configuration is incomplete and previously worked by accident!
setuptools-scm requires setuptools>=61

Suggested workaround if applicable:
 - migrating from the deprecated setup_requires mechanism to pep517/518
   and using a pyproject.toml to declare build dependencies
   which are reliably pre-installed before running the build tools
```